DataGetter tests were added

+++ src/test/java/server/service/DataSetterTest.java
@@ -3,21 +3,23 @@ package server.service;
 import client.Client;
 import client.RequestToServer;
 import java.io.File;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
-import server.Server;
 
 class DataSetterTest {
-    private static final ExecutorService executor = Executors.newFixedThreadPool(3);
+    private static final String DB_FILE_PATH = "src/main/resources/data/db.json";
+    private static final Client client = new Client();
 
     @BeforeAll
-    static void setup() throws InterruptedException {
-        File db = new File("src/main/resources/data/db.json");
-        db.delete();
-        executor.execute(new Server());
-        Thread.sleep(1000);
+    static void setup() {
+        File db = new File(DB_FILE_PATH);
+        if (db.exists()) {
+            db.delete();
+        }
     }
 
     @Test
@@ -27,8 +29,14 @@ class DataSetterTest {
                 .setKey("Person")
                 .setValue("Bob")
                 .build();
-        Client client = new Client();
         client.runClient(setRequest);
+        try {
+            Assertions.assertTrue(
+                    Files.lines(Path.of(DB_FILE_PATH))
+                            .anyMatch(line -> line.contains("\"Person\": \"Bob\"")));
+        } catch (IOException e) {
+            throw new RuntimeException("Can't read file with path: " + DB_FILE_PATH);
+        }
     }
 
     @Test
@@ -36,8 +44,14 @@ class DataSetterTest {
         RequestToServer setRequest = RequestToServer.builder()
                 .setFileName("test2Set.json")
                 .build();
-        Client client = new Client();
         client.runClient(setRequest);
+        try {
+            Assertions.assertTrue(
+                    Files.lines(Path.of(DB_FILE_PATH))
+                            .anyMatch(line -> line.contains("launches\": \"88")));
+        } catch (IOException e) {
+            throw new RuntimeException("Can't read file with path: " + DB_FILE_PATH);
+        }
     }
 
     @Test
@@ -45,8 +59,13 @@ class DataSetterTest {
         RequestToServer setRequest = RequestToServer.builder()
                 .setFileName("test3Set.json")
                 .build();
-        Client client = new Client();
         client.runClient(setRequest);
+        try {
+            Assertions.assertTrue(
+                    Files.lines(Path.of(DB_FILE_PATH))
+                            .anyMatch(line -> line.contains("name\": \"Tesla")));
+        } catch (IOException e) {
+            throw new RuntimeException("Can't read file with path: " + DB_FILE_PATH);
+        }
     }
-
 }
